backend-fastapi/pyproject.toml
[project]
name = "backend-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.1",
    "jwt>=1.4.0",
    "langchain>=1.2.17",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.2",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.10",
    "motor>=3.7.1",
    "pydantic>=2.13.3",
    "pyjwt>=2.12.1",
    "pymongo>=4.17.0",
    "python-dotenv>=1.2.2",
    "pytz>=2026.1.post1",
    "tavily-python>=0.7.24",
    "uvicorn>=0.46.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
